+++ python_ta/checkers/forbidden_import_checker.py
@@ -1,3 +1,5 @@
+"""Checker or use of forbidden imports.
+"""
 import os
 
 from astroid import nodes
@@ -7,6 +9,9 @@ from pylint.lint import PyLinter
 
 
 class ForbiddenImportChecker(BaseChecker):
+    """A checker class to report on the disallowed imports in the file.
+    Use options to specify the import modules that are allowed and the extra import modules."""
+
     name = "forbidden_import"
     msgs = {
         "E9999": (
@@ -35,9 +40,6 @@ class ForbiddenImportChecker(BaseChecker):
             },
         ),
     )
-
-    # this is important so that your checker is executed before others
-    priority = -1
 
     @only_required_for_messages("forbidden-import")
     def visit_import(self, node: nodes.Import) -> None:
@@ -83,7 +85,6 @@ class ForbiddenImportChecker(BaseChecker):
                         and node.args[0].value not in self.get_local_files()
                     ):
                         args = (node.args[0].value, node.lineno)
-                        # add the message
                         self.add_message("forbidden-import", node=node, args=args)
 
     def get_local_files(self) -> list:
@@ -102,5 +103,5 @@ class ForbiddenImportChecker(BaseChecker):
 
 
 def register(linter: PyLinter) -> None:
-    """required method to auto register this checker"""
+    """Required method to auto register this checker"""
     linter.register_checker(ForbiddenImportChecker(linter))